# Performance Improvements for string().guid()