--- a/lib/types/string/index.js
+++ b/lib/types/string/index.js
@@ -338,9 +338,12 @@ internals.String = class extends Any {
             'uuidv5': '5'
         };
 
-        const versions = [];
+        let versionNumbers = [];
+        const versions = {};
+        let checkVersion = false;
 
         if (guidOptions && guidOptions.version) {
+            checkVersion = true;
             if (!Array.isArray(guidOptions.version)) {
                 guidOptions.version = [guidOptions.version];
             }
@@ -351,43 +354,30 @@ internals.String = class extends Any {
                 let version = guidOptions.version[i];
                 Hoek.assert(typeof version === 'string', 'version at position ' + i + ' must be a string');
                 version = version.toLowerCase();
-                Hoek.assert(uuids[version], 'version at position ' + i + ' must be one of ' + Object.keys(uuids).join(', '));
-                Hoek.assert(versions.indexOf(version) === -1, 'version at position ' + i + ' must not be a duplicate.');
-                versions.push(version);
+                const versionNumber = uuids[version];
+                Hoek.assert(versionNumber, 'version at position ' + i + ' must be one of ' + Object.keys(uuids).join(', '));
+                Hoek.assert(!(versionNumber in versions), 'version at position ' + i + ' must not be a duplicate.');
+
+                versionNumbers.push(versionNumber);
+                versions[versionNumber] = '';
             }
+
+            versionNumbers = versionNumbers.join('');
         }
 
-        const regex = /^([\[{\(]?)([0-9A-F]{8})([:-]?)([0-9A-F]{4})([:-]?)([0-9A-F]{4})([:-]?)([0-9A-F]{4})([:-]?)([0-9A-F]{12})([\]}\)]?)$/i;
+        const guidRegex = new RegExp(`^([\\[{\\(]?)[0-9A-F]{8}([:-]?)[0-9A-F]{4}\\2?[${checkVersion ? versionNumbers : '0-9A-F'}][0-9A-F]{3}\\2?[${checkVersion ? '89AB' : '0-9A-F'}][0-9A-F]{3}\\2?[0-9A-F]{12}([\\]}\\)]?)$`, 'i');
 
         return this._test('guid', guidOptions, function (value, state, options) {
 
-            const results = regex.exec(value);
+            const results = guidRegex.exec(value);
 
             if (!results) {
                 return this.createError('string.guid', { value }, state, options);
             }
 
             // Matching braces
-            if (brackets[results[1]] !== results[11]) {
+            if (brackets[results[1]] !== results[results.length - 1]) {
                 return this.createError('string.guid', { value }, state, options);
-            }
-
-            // Matching separators
-            if (results[3] !== results[5] || results[3] !== results[7] || results[3] !== results[9]) {
-                return this.createError('string.guid', { value }, state, options);
-            }
-
-            // Specific UUID versions
-            if (versions.length) {
-                const validVersions = versions.some((uuidVersion) => {
-
-                    return results[6][0] === uuids[uuidVersion];
-                });
-
-                // Valid version and 89AB check
-                if (!(validVersions && /[89AB]/i.test(results[8][0]))) {
-                    return this.createError('string.guid', { value }, state, options);
-                }
             }
 
             return value;