--- a/lib/types/string/index.js
+++ b/lib/types/string/index.js
@@ -338,12 +338,10 @@ internals.String = class extends Any {
             'uuidv5': '5'
         };
 
-        let versionNumbers = [];
+        let versionNumbers = '';
         const versions = {};
-        let checkVersion = false;
 
         if (guidOptions && guidOptions.version) {
-            checkVersion = true;
             if (!Array.isArray(guidOptions.version)) {
                 guidOptions.version = [guidOptions.version];
             }
@@ -358,14 +356,12 @@ internals.String = class extends Any {
                 Hoek.assert(versionNumber, 'version at position ' + i + ' must be one of ' + Object.keys(uuids).join(', '));
                 Hoek.assert(!(versionNumber in versions), 'version at position ' + i + ' must not be a duplicate.');
 
-                versionNumbers.push(versionNumber);
+                versionNumbers += versionNumber;
                 versions[versionNumber] = '';
             }
-
-            versionNumbers = versionNumbers.join('');
         }
 
-        const guidRegex = new RegExp(`^([\\[{\\(]?)[0-9A-F]{8}([:-]?)[0-9A-F]{4}\\2?[${checkVersion ? versionNumbers : '0-9A-F'}][0-9A-F]{3}\\2?[${checkVersion ? '89AB' : '0-9A-F'}][0-9A-F]{3}\\2?[0-9A-F]{12}([\\]}\\)]?)$`, 'i');
+        const guidRegex = new RegExp(`^([\\[{\\(]?)[0-9A-F]{8}([:-]?)[0-9A-F]{4}\\2?[${versionNumbers || '0-9A-F'}][0-9A-F]{3}\\2?[${versionNumbers ? '89AB' : '0-9A-F'}][0-9A-F]{3}\\2?[0-9A-F]{12}([\\]}\\)]?)$`, 'i');
 
         return this._test('guid', guidOptions, function (value, state, options) {