--- a/lib/types/string/index.js
+++ b/lib/types/string/index.js
@@ -15,7 +15,17 @@ const Ip = require('./ip');
 
 const internals = {
     uriRegex: Uri.createUriRegex(),
-    ipRegex: Ip.createIpRegex(['ipv4', 'ipv6', 'ipvfuture'], 'optional')
+    ipRegex: Ip.createIpRegex(['ipv4', 'ipv6', 'ipvfuture'], 'optional'),
+    guidBrackets: {
+        '{': '}', '[': ']', '(': ')', '': ''
+    },
+    guidVersions: {
+        uuidv1: '1',
+        uuidv2: '2',
+        uuidv3: '3',
+        uuidv4: '4',
+        uuidv5: '5'
+    }
 };
 
 internals.String = class extends Any {
@@ -326,20 +336,7 @@ internals.String = class extends Any {
 
     guid(guidOptions) {
 
-        const brackets = {
-            '{': '}', '[': ']', '(': ')', '': ''
-        };
-
-        const uuids = {
-            'uuidv1': '1',
-            'uuidv2': '2',
-            'uuidv3': '3',
-            'uuidv4': '4',
-            'uuidv5': '5'
-        };
-
         let versionNumbers = '';
-        const versions = {};
 
         if (guidOptions && guidOptions.version) {
             if (!Array.isArray(guidOptions.version)) {
@@ -347,17 +344,18 @@ internals.String = class extends Any {
             }
 
             Hoek.assert(guidOptions.version.length >= 1, 'version must have at least 1 valid version specified');
+            const versions = new Set();
 
             for (let i = 0; i < guidOptions.version.length; ++i) {
                 let version = guidOptions.version[i];
                 Hoek.assert(typeof version === 'string', 'version at position ' + i + ' must be a string');
                 version = version.toLowerCase();
-                const versionNumber = uuids[version];
-                Hoek.assert(versionNumber, 'version at position ' + i + ' must be one of ' + Object.keys(uuids).join(', '));
-                Hoek.assert(!(versionNumber in versions), 'version at position ' + i + ' must not be a duplicate.');
+                const versionNumber = internals.guidVersions[version];
+                Hoek.assert(versionNumber, 'version at position ' + i + ' must be one of ' + Object.keys(internals.guidVersions).join(', '));
+                Hoek.assert(!(versions.has(versionNumber)), 'version at position ' + i + ' must not be a duplicate.');
 
                 versionNumbers += versionNumber;
-                versions[versionNumber] = '';
+                versions.add(versionNumber);
             }
         }
 
@@ -372,7 +370,7 @@ internals.String = class extends Any {
             }
 
             // Matching braces
-            if (brackets[results[1]] !== results[results.length - 1]) {
+            if (internals.guidBrackets[results[1]] !== results[results.length - 1]) {
                 return this.createError('string.guid', { value }, state, options);
             }