This PR includes some performance improvements to the run-time validation of the `string().guid()` validation type. 

#### How Was This Determined?
I ran a [series of benchmarks](https://github.com/DavidTPate/joi-bench/blob/dabe172db0e2aa81f43ea86b0381a2d4f11e7e02/benchmarks/string.js#L135-L158) for some common cases of GUID/UUIDs and then made tweaks to improve the performance.

**Schemas**
```js
const guidSchema = stringSchema.guid();
const guidVersionsSchema = stringSchema.guid({
   version: ['uuidv1', 'uuidv2', 'uuidv3', 'uuidv4']
});
```

**Benchmarks**
```js
// validate-schema#guid
guidSchema.validate('D1A5279D-B27D-4CD4-A05E-EFDD53D08E8D');

// validate-schema#guid-wrapped
guidSchema.validate('{D1A5279D-B27D-4CD4-A05E-EFDD53D08E8D}');

// validate-schema#guid-versions
guidVersionsSchema.validate('D1A5279D-B27D-4CD4-A05E-EFDD53D08E8D');

// validate-schema#guid-versions-wrapped
guidVersionsSchema.validate('{D1A5279D-B27D-4CD4-A05E-EFDD53D08E8D}');
```

#### Original Benchmarks
Unmodified I received the following benchmarks:

```
string
  validate-schema#guid-versions         x 353,757 ops/sec
  validate-schema#guid-versions-wrapped x 370,182 ops/sec
  validate-schema#guid                  x 406,565 ops/sec
  validate-schema#guid-wrapped          x 423,820 ops/sec
```

From there I deferred more of the work to the Regular Expression but kept the version check within an `if` statement, which brought me the following results:

```
string
  validate-schema#guid                  x 428,658 ops/sec
  validate-schema#guid-versions         x 439,570 ops/sec
  validate-schema#guid-versions-wrapped x 451,946 ops/sec
  validate-schema#guid-wrapped          x 458,762 ops/sec
```

From there I went and pulled the version & `89AB` checks along with utilizing back references within the Regular Expression for divider checks. Additionally, I simplified the Regular Expression to eliminate capture & non-capture groups where possible. This lead to the follow (current) benchmarks:

```
string
  validate-schema#guid-versions         x 449,913 ops/sec
  validate-schema#guid                  x 458,897 ops/sec
  validate-schema#guid-versions-wrapped x 472,751 ops/sec
  validate-schema#guid-wrapped          x 473,101 ops/sec
```

> **Note**: These benchmarks were run on my machine, so if you run them expect the numbers themselves to vary (also by Node version & platform), but they should stay around the same magnitude of difference.