@Marsup 2017-06-05
According to my tests, it seems to be an improvement on node 6, but it's very much the opposite for node 8, not sure what to do here. I'd tend to favor the future version. What do you get on your machine ?