@DavidTPate 2017-06-05
That's very interesting as I noticed improvements on Node 8 as well.

Node 6.10.3 (Fedora 24 - 64 bit)
```
string
  validate-schema#guid-versions         x 372,592 ops/sec
  validate-schema#guid-versions-wrapped x 386,855 ops/sec
  validate-schema#guid                  x 427,795 ops/sec
  validate-schema#guid-wrapped          x 444,966 ops/sec
```

Node 6.10.3 (Fedora 24 - 64 bit) - New Code
```
string
  validate-schema#guid                  x 491,424 ops/sec
  validate-schema#guid-versions         x 502,937 ops/sec
  validate-schema#guid-versions-wrapped x 517,636 ops/sec
  validate-schema#guid-wrapped          x 529,663 ops/sec
```

Node 8.0.0 (Fedora 24 - 64 bit)
```
string
  validate-schema#guid-versions-wrapped x 388,319 ops/sec
  validate-schema#guid-versions         x 398,400 ops/sec
  validate-schema#guid-wrapped          x 450,996 ops/sec
  validate-schema#guid                  x 473,374 ops/sec
```

Node 8.0.0 (Fedora 24 - 64 bit) - New Code
```
string
  validate-schema#guid                  x 451,330 ops/sec
  validate-schema#guid-wrapped          x 454,979 ops/sec
  validate-schema#guid-versions-wrapped x 471,269 ops/sec
  validate-schema#guid-versions         x 499,805 ops/sec
```