@DavidTPate 2017-06-05
Forgot to follow up on this as well:

> I'd tend to favor the future version.

Definitely agree.