@Marsup 2017-06-13
Confirmed improvement on 8.1 ¯\\\_(ツ)\_/¯

I'll review soon.