@DavidTPate 2017-06-14
So I tried a few different versions. Ultimately the Set was the most consistent and fastest, results from the 3 different tests are below:

#### Object In
string
  validate-schema#guid-versions-wrapped x 433,471 ops/sec
  validate-schema#guid-wrapped          x 443,962 ops/sec
  validate-schema#guid-versions         x 458,346 ops/sec
  validate-schema#guid                  x 464,135 ops/sec

#### Set
string
  validate-schema#guid-wrapped          x 461,747 ops/sec
  validate-schema#guid-versions-wrapped x 468,894 ops/sec
  validate-schema#guid-versions         x 476,740 ops/sec
  validate-schema#guid                  x 481,809 ops/sec


#### hasOwnProperty
string
  validate-schema#guid-wrapped          x 417,693 ops/sec
  validate-schema#guid-versions-wrapped x 418,105 ops/sec
  validate-schema#guid-versions         x 434,678 ops/sec
  validate-schema#guid                  x 436,262 ops/sec