@Marsup 2017-06-14
Perfect, I'll be waiting for your update with a Set then.

FYI I have noticed the same kind of improvement using Maps instead of objects as hash, I might be switching joi to that where it makes sense at some point.